ci: expand golangci-lint set and fix findings
Enabled linters: errorlint, gocritic, gosec, misspell, nilerr, revive,
sqlclosecheck, unconvert, bodyclose. Swapped goimports for gci with
explicit stdlib/default/local section order. Made baseline linter list
explicit (default: none).

Code fixes:
- errorlint: replace err != io.EOF with errors.Is; switch %v to %w
  in fmt.Errorf (viewer.go)
- gocritic ifElseChain -> switch: app.go x2, view.go x3 (including
  nested status dispatch), format.go, file_viewer/viewer.go
- gocritic assignOp: x += y cleanups in terminal_osc, file_list,
  archive, database_table_content
- gocritic elseif: flatten else-if in viewer_test.go
- gocritic exitAfterDefer: drop defer on debug log; explicit Close()
  on both exit paths in main.go
- revive indent-error-flow: drop else-after-return in update.go and
  file_viewer/viewer.go
- revive redefines-builtin-id: delete local func max (go 1.21 builtin)
- unconvert: drop redundant uint8() casts in theme.go and int() cast
  in update.go

Security:
- gosec G301: tighten ~/.config/simtool dir perms 0755 -> 0700
- gosec G201 (SQL identifier injection): add quoteSQLiteIdentifier
  helper and rewrite getTableRowCount/getTableColumns/getTableSample/
  ReadTableData to use identifier escaping with strconv.Itoa for
  numeric params. sqlite cannot parameterize identifiers.

Config-level linter silencing (CLAUDE.md forbids per-line pragmas):
- revive: disable exported, package-comments, unused-parameter,
  blank-imports (noise, not bug detection)
- gosec: disable G115 (syscall int casts), G201/G202 (identifier
  safety handled by quoteSQLiteIdentifier), G204/G702 (exec.Command
  isn't shell-interpreted), G304/G703 (viewer intentionally reads
  user-selected paths). Each exclusion has a reason comment.

No test changes. All 5 packages still pass.

+++ internal/config/theme.go
@@ -244,10 +244,7 @@ func colorToHex(c chroma.Colour) string {
 	if !c.IsSet() {
 		return ""
 	}
-	return fmt.Sprintf("#%02x%02x%02x",
-		uint8(c.Red()),
-		uint8(c.Green()),
-		uint8(c.Blue()))
+	return fmt.Sprintf("#%02x%02x%02x", c.Red(), c.Green(), c.Blue())
 }
 
 // isLightTheme determines if a theme is light based on background brightness
